+++ test/deploy.test.js
@@ -61,7 +61,7 @@ describe('Deploy Test', () => {
     builder._deployers.wsk.init = () => {};
 
     await assert.rejects(builder.run(), /Openwhisk target needs --wsk-host, --wsk-auth and --wsk-namespace/);
-  });
+  }).timeout(5000);
 
   it('reports error configured namespace does not match wsk namespace', async () => {
     await fse.copy(path.resolve(__dirname, 'fixtures', 'web-action'), testRoot);
@@ -156,6 +156,30 @@ describe('Deploy Test', () => {
     const out = builder.cfg.log.output;
     assert.ok(out.indexOf('requesting: https://example.com/api/v1/web/foobar/default/simple-project/foo') > 0);
     assert.ok(out.indexOf('Location: https://example.com/') > 0);
+  }).timeout(10000);
+
+  it('tests can send headers', async () => {
+    await fse.copy(path.resolve(__dirname, 'fixtures', 'web-action'), testRoot);
+
+    nock(process.env.WSK_APIHOST)
+      .get('/api/v1/web/foobar/default/simple-project/foo')
+      .matchHeader('x-foo-bar', 'test')
+      .reply(200, 'ok');
+    process.chdir(testRoot); // need to change .cwd() for yargs to pickup `wsk` in package.json
+    const builder = new CLI()
+      .prepare([
+        '--target', 'wsk',
+        '--verbose',
+        '--no-build',
+        '--test', '/foo',
+        '--test-headers', 'x-foo-bar=test',
+        '--directory', testRoot,
+      ]);
+    builder.cfg._logger = new TestLogger();
+
+    await builder.run();
+    const out = builder.cfg.log.output;
+    assert.ok(out.indexOf('requesting: https://example.com/api/v1/web/foobar/default/simple-project/foo') > 0);
   }).timeout(10000);
 
   it('test can retry with 404', async () => {